Bug Fix 1886: Refactored ImageFusionView
